requirements.txt
langchain-anthropic>=0.3.0
langgraph>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
python-dotenv>=1.0.0
httpx>=0.27.0
biopython>=1.84
chromadb>=0.5.0
voyageai>=0.3.0
fpdf2>=2.7.0
fastmcp>=2.0.0
beautifulsoup4>=4.12.0
